refactor(plugins/server): fix imports
